Update diversitynow.scss
With a zIndex of 999 it covers up the navigation hover over text "Code of Conduct".  I tested it with a setting of 1 and scrolled through the page and the diversity banner behaved as expected.  Setting it at 10 as that also worked can could allow for other layers in the future.